feat: unify format convert skill & upgrade downloader with Motrix UI and Linux CLI
