+++ skills/downloader/__init__.py
@@ -1,1354 +1,29 @@
-import os
-import sys
-import json
-import time
-import threading
-import hashlib
-import urllib.request
-import urllib.parse
-import http.server
-import socket
-import re
-import shutil
-import logging
-import zipfile
-from typing import Dict, List, Any
-
-# Ensure BeautifulSoup4 and requests are available
-try:
-    from bs4 import BeautifulSoup
-except ImportError:
-    BeautifulSoup = None
-
-try:
-    import requests
-except ImportError:
-    requests = None
-
-logger = logging.getLogger("DownloaderSkill")
-
-# Path Configuration
-PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-SKILL_DIR = os.path.join(PROJECT_ROOT, "skills", "downloader")
-DATA_DIR = os.path.join(SKILL_DIR, "data")
-os.makedirs(DATA_DIR, exist_ok=True)
-
-CONFIG_PATH = os.path.join(DATA_DIR, "tasks.json")
-OLD_CONFIG_PATH = os.path.join(SKILL_DIR, "tasks.json")
-
-# Migrate old tasks.json if it exists and the new one doesn't
-if os.path.exists(OLD_CONFIG_PATH) and not os.path.exists(CONFIG_PATH):
-    try:
-        shutil.move(OLD_CONFIG_PATH, CONFIG_PATH)
-    except Exception:
-        pass
-
-# Local config & standalone settings helper
-LOCAL_CONFIG_PATH = os.path.join(DATA_DIR, "config.json")
-IS_STANDALONE = False
-
-def get_standalone_status() -> bool:
-    global IS_STANDALONE
-    return IS_STANDALONE or os.environ.get("BUTLER_DOWNLOADER_STANDALONE") == "1"
-
-def load_local_config() -> Dict[str, Any]:
-    if not os.path.exists(LOCAL_CONFIG_PATH):
-        return {}
-    try:
-        with open(LOCAL_CONFIG_PATH, "r", encoding="utf-8") as f:
-            return json.load(f)
-    except Exception:
-        return {}
-
-def save_local_config(config: Dict[str, Any]):
-    try:
-        with open(LOCAL_CONFIG_PATH, "w", encoding="utf-8") as f:
-            json.dump(config, f, indent=4, ensure_ascii=False)
-    except Exception:
-        pass
-
-def get_downloads_dir() -> str:
-    cfg = load_local_config()
-    custom_path = cfg.get("download_path")
-    if custom_path:
-        # Standardize path
-        custom_path = os.path.expanduser(custom_path)
-        if not os.path.isabs(custom_path):
-            custom_path = os.path.abspath(os.path.join(PROJECT_ROOT, custom_path))
-        os.makedirs(custom_path, exist_ok=True)
-        return custom_path
-
-    # Defaults
-    if get_standalone_status():
-        path = os.path.join(SKILL_DIR, "downloads")
-    else:
-        path = os.path.join(PROJECT_ROOT, "data", "downloads")
-    os.makedirs(path, exist_ok=True)
-    return path
-
-# Legacy compatibility global variable (evaluated once at load)
-DOWNLOADS_DIR = get_downloads_dir()
-
-# Global variables
-streaming_server_port = 8329
-streaming_thread = None
-scheduler_thread = None
-scheduler_running = False
-
-# Active downloading instances: task_id -> { "thread": Thread, "stop_flag": Event, "manager": chunk_downloader }
-ACTIVE_DOWNLOADS = {}
-active_lock = threading.Lock()
-
-class SafeHTTPRangeHandler(http.server.BaseHTTPRequestHandler):
-    """
-    A high-fidelity Range-supporting HTTP Server for local streaming.
-    Supports CORS, Range bytes requests (HTTP 206), and Head requests.
-    This enables real-time HTML5 "Play during Download" (边下边播) for media files.
-    """
-    def log_message(self, format, *args):
-        # Suppress noisy standard log output
-        pass
-
-    def do_OPTIONS(self):
-        self.send_response(200)
-        self.send_header("Access-Control-Allow-Origin", "*")
-        self.send_header("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
-        self.send_header("Access-Control-Allow-Headers", "Range, Content-Type")
-        self.end_headers()
-
-    def do_GET(self):
-        parsed = urllib.parse.urlparse(self.path)
-        if parsed.path in ["/ui/", "/ui/index.html"]:
-            ui_file_path = os.path.join(PROJECT_ROOT, "skills", "downloader", "ui", "index.html")
-            if os.path.exists(ui_file_path):
-                self.send_response(200)
-                self.send_header("Content-Type", "text/html; charset=utf-8")
-                self.send_header("Access-Control-Allow-Origin", "*")
-                self.end_headers()
-                with open(ui_file_path, "rb") as f:
-                    self.wfile.write(f.read())
-            else:
-                self.send_error(404, "UI File Not Found")
-            return
-
-        self.handle_request()
-
-    def do_POST(self):
-        parsed = urllib.parse.urlparse(self.path)
-        if parsed.path.startswith("/api/"):
-            action = parsed.path[5:] # strip "/api/"
-            content_length = int(self.headers.get("Content-Length", 0))
-            body = self.rfile.read(content_length).decode("utf-8") if content_length > 0 else "{}"
-            try:
-                params = json.loads(body)
-            except Exception:
-                params = {}
-
-            # Execute backend action
-            result = handle_request(action, **params)
-
-            self.send_response(200)
-            self.send_header("Content-Type", "application/json; charset=utf-8")
-            self.send_header("Access-Control-Allow-Origin", "*")
-            self.end_headers()
-            self.wfile.write(json.dumps(result, ensure_ascii=False).encode("utf-8"))
-            return
-
-        self.send_error(404, "Not Found")
-
-    def do_HEAD(self):
-        self.handle_request(head_only=True)
-
-    def handle_request(self, head_only=False):
-        parsed = urllib.parse.urlparse(self.path)
-        if parsed.path != "/stream":
-            self.send_error(404, "Not Found")
-            return
-
-        query = urllib.parse.parse_qs(parsed.query)
-        task_id = query.get("task_id", [None])[0]
-        if not task_id:
-            self.send_error(400, "Missing task_id")
-            return
-
-        # Load task to find file path
-        tasks = load_tasks()
-        task = tasks.get(task_id)
-        if not task:
-            self.send_error(404, "Task not found")
-            return
-
-        file_path = task.get("file_path")
-        if not file_path or not os.path.exists(file_path):
-            # If completed file is missing, check if temporary/part files exist
-            # Or use partial downloaded file for streaming
-            part_path = file_path + ".part" if file_path else None
-            if part_path and os.path.exists(part_path):
-                file_path = part_path
-            else:
-                self.send_error(404, f"File does not exist on disk: {file_path}")
-                return
-
-        file_size = os.path.getsize(file_path)
-        content_type = "video/mp4" # Default media stream
-        if file_path.lower().endswith(".mp3"):
-            content_type = "audio/mpeg"
-        elif file_path.lower().endswith(".mkv"):
-            content_type = "video/x-matroska"
-
-        # Check Range request header
-        range_header = self.headers.get("Range")
-        start, end = 0, file_size - 1
-
-        is_partial = False
-        if range_header:
-            match = re.match(r"bytes=(\d+)-(\d*)", range_header)
-            if match:
-                is_partial = True
-                start = int(match.group(1))
-                if match.group(2):
-                    end = int(match.group(2))
-
-        # Boundaries validation
-        if start >= file_size:
-            self.send_response(416, "Range Not Satisfiable")
-            self.send_header("Content-Range", f"bytes */{file_size}")
-            self.end_headers()
-            return
-
-        if end >= file_size:
-            end = file_size - 1
-
-        content_length = end - start + 1
-
-        # Send response headers
-        if is_partial:
-            self.send_response(206, "Partial Content")
-            self.send_header("Content-Range", f"bytes {start}-{end}/{file_size}")
-        else:
-            self.send_response(200, "OK")
-
-        self.send_header("Content-Type", content_type)
-        self.send_header("Content-Length", str(content_length))
-        self.send_header("Accept-Ranges", "bytes")
-        self.send_header("Access-Control-Allow-Origin", "*")
-        self.end_headers()
-
-        if head_only:
-            return
-
-        # Read and stream the file chunk by chunk
-        try:
-            with open(file_path, "rb") as f:
-                f.seek(start)
-                remaining = content_length
-                chunk_size = 1024 * 64
-                while remaining > 0:
-                    read_len = min(chunk_size, remaining)
-                    data = f.read(read_len)
-                    if not data:
-                        break
-                    self.wfile.write(data)
-                    remaining -= len(data)
-        except Exception as e:
-            # Socket connection aborted by player is normal during streaming seeking
-            pass
-
-
-def start_streaming_server():
-    """Launches the Streaming HTTP server in a background thread."""
-    global streaming_server_port, streaming_thread
-    server_address = ("", streaming_server_port)
-    for port in range(8329, 8340):
-        try:
-            httpd = http.server.ThreadingHTTPServer(("", port), SafeHTTPRangeHandler)
-            streaming_server_port = port
-            logger.info(f"Stream Streaming Server launched on port {port}")
-
-            def run_server():
-                try:
-                    httpd.serve_forever()
-                except Exception:
-                    pass
-
-            streaming_thread = threading.Thread(target=run_server, daemon=True)
-            streaming_thread.start()
-            break
-        except OSError:
-            continue
-
-
-def load_tasks() -> Dict[str, Any]:
-    """Loads download tasks from tasks.json safely."""
-    if not os.path.exists(CONFIG_PATH):
-        return {}
-    try:
-        with open(CONFIG_PATH, "r", encoding="utf-8") as f:
-            return json.load(f)
-    except Exception as e:
-        logger.error(f"Failed to load tasks: {e}")
-        return {}
-
-
-def save_tasks(tasks: Dict[str, Any]):
-    """Persists tasks dictionary to tasks.json."""
-    try:
-        with open(CONFIG_PATH, "w", encoding="utf-8") as f:
-            json.dump(tasks, f, indent=4, ensure_ascii=False)
-    except Exception as e:
-        logger.error(f"Failed to save tasks: {e}")
-
-
-def calculate_eta(downloaded, total, speed_bytes) -> str:
-    """Helper to format ETA into friendly terms."""
-    if not speed_bytes or speed_bytes <= 0 or total <= 0:
-        return "未知"
-    remaining_bytes = total - downloaded
-    seconds = int(remaining_bytes / speed_bytes)
-    if seconds < 60:
-        return f"{seconds}秒"
-    minutes = seconds // 60
-    if minutes < 60:
-        return f"{minutes}分{seconds % 60}秒"
-    hours = minutes // 60
-    return f"{hours}时{minutes % 60}分"
-
-
-def format_bytes(bytes_count) -> str:
-    """Helper to format byte count into human-readable units."""
-    if bytes_count < 1024:
-        return f"{bytes_count} B"
-    elif bytes_count < 1024 * 1024:
-        return f"{bytes_count / 1024:.1f} KB"
-    elif bytes_count < 1024 * 1024 * 1024:
-        return f"{bytes_count / (1024 * 1024):.1f} MB"
-    else:
-        return f"{bytes_count / (1024 * 1024 * 1024):.1f} GB"
-
-
-class SegmentedDownloader:
-    """
-    Core Multi-threaded HTTP/HTTPS segment downloader.
-    Supports range downloads, resuming, speed limit throttling, and smart retries.
-    """
-    def __init__(self, task_id: str, url: str, file_path: str, max_workers: int = 16, chunk_size: int = 1024 * 64):
-        self.task_id = task_id
-        self.url = url
-        self.file_path = file_path
-        self.max_workers = max_workers
-        self.chunk_size = chunk_size
-
-        self.total_size = 0
-        self.downloaded_size = 0
-        self.supports_range = False
-        self.stop_event = threading.Event()
-        self.error_message = ""
-
-        # Real-time stats
-        self.start_time = 0.0
-        self.speed = 0.0
-
-        # Throttling
-        self.speed_limit = 0 # 0 means unlimited, bytes/second
-
-    def precheck(self):
-        """Sends a HEAD request to fetch content size and range support."""
-        try:
-            req = urllib.request.Request(self.url, method="HEAD")
-            req.add_header("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36")
-            with urllib.request.urlopen(req, timeout=10) as response:
-                self.total_size = int(response.headers.get("Content-Length", 0))
-                accept_ranges = response.headers.get("Accept-Ranges", "")
-                self.supports_range = "bytes" in accept_ranges or response.headers.get("Content-Range") is not None
-        except Exception:
-            # Fallback to GET with dynamic range checking
-            try:
-                req = urllib.request.Request(self.url)
-                req.add_header("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36")
-                req.add_header("Range", "bytes=0-0")
-                with urllib.request.urlopen(req, timeout=10) as response:
-                    self.supports_range = (response.status == 206)
-                    content_range = response.headers.get("Content-Range", "")
-                    if content_range and "/" in content_range:
-                        self.total_size = int(content_range.split("/")[-1])
-            except Exception as e:
-                logger.error(f"Precheck failed: {e}")
-                self.supports_range = False
-
-    def download_segment(self, seg_index: int, start: int, end: int, part_path: str, progress_list: List[int]):
-        """Downloads a single segment chunk, supporting 3 smart retries."""
-        retries = 3
-        while retries > 0 and not self.stop_event.is_set():
-            try:
-                # Calculate resume position of this segment if part file exists
-                current_start = start
-                mode = "wb"
-                if os.path.exists(part_path):
-                    current_start += os.path.getsize(part_path)
-                    mode = "ab"
-                    if current_start > end:
-                        # Already completed segment
-                        progress_list[seg_index] = end - start + 1
-                        return
-
-                if current_start > end:
-                    return
-
-                req = urllib.request.Request(self.url)
-                req.add_header("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36")
-                req.add_header("Range", f"bytes={current_start}-{end}")
-
-                with urllib.request.urlopen(req, timeout=15) as conn, open(part_path, mode) as f:
-                    while not self.stop_event.is_set():
-                        # Throttling calculation
-                        t0 = time.time()
-                        data = conn.read(self.chunk_size)
-                        if not data:
-                            break
-
-                        f.write(data)
-                        chunk_len = len(data)
-                        self.downloaded_size += chunk_len
-                        progress_list[seg_index] += chunk_len
-
-                        # Update speed statistics periodically outside thread or sleep to limit speed
-                        if self.speed_limit > 0:
-                            # Simple pacing logic: segment share of speed limit
-                            seg_speed_limit = self.speed_limit / min(self.max_workers, 4) # share limit
-                            elapsed = time.time() - t0
-                            expected_time = chunk_len / seg_speed_limit
-                            if elapsed < expected_time:
-                                time.sleep(expected_time - elapsed)
-
-                # Completed segment successfully
-                return
-            except Exception as e:
-                retries -= 1
-                logger.warning(f"Segment {seg_index} failed, retrying ({3-retries}/3). Error: {e}")
-                time.sleep(1)
-
-        if retries == 0 and not self.stop_event.is_set():
-            self.error_message = f"分块 {seg_index} 下载重试失败"
-            self.stop_event.set()
-
-    def download_sequential(self):
-        """Downloads the file sequentially (fallback mode for servers that do not support ranges)."""
-        retries = 3
-        part_path = self.file_path + ".part"
-        while retries > 0 and not self.stop_event.is_set():
-            try:
-                current_start = 0
-                mode = "wb"
-                if os.path.exists(part_path):
-                    current_start = os.path.getsize(part_path)
-                    mode = "ab"
-
-                req = urllib.request.Request(self.url)
-                req.add_header("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36")
-                if current_start > 0:
-                    req.add_header("Range", f"bytes={current_start}-")
-
-                with urllib.request.urlopen(req, timeout=20) as conn, open(part_path, mode) as f:
-                    while not self.stop_event.is_set():
-                        t0 = time.time()
-                        data = conn.read(self.chunk_size)
-                        if not data:
-                            break
-
-                        f.write(data)
-                        chunk_len = len(data)
-                        self.downloaded_size += chunk_len
-
-                        # Limit speed
-                        if self.speed_limit > 0:
-                            elapsed = time.time() - t0
-                            expected_time = chunk_len / self.speed_limit
-                            if elapsed < expected_time:
-                                time.sleep(expected_time - elapsed)
-
-                # Success
-                if not self.stop_event.is_set():
-                    shutil.move(part_path, self.file_path)
-                return
-            except Exception as e:
-                retries -= 1
-                logger.warning(f"Sequential download failed: {e}. Retrying...")
-                time.sleep(2)
-
-        if retries == 0:
-            self.error_message = "单线程顺序下载失败，请检查网络"
-            self.stop_event.set()
-
-    def run(self):
-        """Main orchestrator for segmented download."""
-        self.start_time = time.time()
-        self.precheck()
-
-        # Update tasks size
-        tasks = load_tasks()
-        if self.task_id in tasks:
-            tasks[self.task_id]["total_size"] = self.total_size
-            tasks[self.task_id]["status"] = "downloading"
-            save_tasks(tasks)
-
-        # Monitor thread to calculate speed and save progress
-        def monitor_loop():
-            last_bytes = 0
-            last_time = time.time()
-            while not self.stop_event.is_set() and self.downloaded_size < (self.total_size or 1):
-                time.sleep(0.8)
-                now = time.time()
-                dt = now - last_time
-                if dt <= 0:
-                    continue
-                db = self.downloaded_size - last_bytes
-                self.speed = db / dt
-
-                # Update stats inside database/tasks.json
-                cur_tasks = load_tasks()
-                if self.task_id in cur_tasks:
-                    cur_tasks[self.task_id]["downloaded_size"] = self.downloaded_size
-                    progress_val = int((self.downloaded_size / self.total_size) * 100) if self.total_size > 0 else 0
-                    cur_tasks[self.task_id]["progress"] = progress_val
-                    cur_tasks[self.task_id]["speed"] = format_bytes(int(self.speed)) + "/s"
-                    cur_tasks[self.task_id]["eta"] = calculate_eta(self.downloaded_size, self.total_size, self.speed)
-                    save_tasks(cur_tasks)
-
-                last_bytes = self.downloaded_size
-                last_time = now
-
-        m_thread = threading.Thread(target=monitor_loop, daemon=True)
-        m_thread.start()
-
-        if self.supports_range and self.total_size > 0:
-            # Segmented multi-threading
-            seg_size = self.total_size // self.max_workers
-            threads = []
-            part_paths = []
-            progress_list = [0] * self.max_workers
-
-            for i in range(self.max_workers):
-                start = i * seg_size
-                end = self.total_size - 1 if i == self.max_workers - 1 else (i + 1) * seg_size - 1
-                part_path = f"{self.file_path}.part_{i}"
-                part_paths.append(part_path)
-
-                t = threading.Thread(
-                    target=self.download_segment,
-                    args=(i, start, end, part_path, progress_list),
-                    daemon=True
-                )
-                threads.append(t)
-                t.start()
-
-            for t in threads:
-                t.join()
-
-            # Merge segments
-            if not self.stop_event.is_set():
-                try:
-                    with open(self.file_path, "wb") as outfile:
-                        for p_path in part_paths:
-                            with open(p_path, "rb") as infile:
-                                shutil.copyfileobj(infile, outfile)
-                            os.remove(p_path)
-                except Exception as e:
-                    self.error_message = f"分块合并失败: {e}"
-                    self.stop_event.set()
-        else:
-            # Sequential sequential download
-            self.download_sequential()
-
-        # Wrap up task execution
-        self.stop_event.set()
-        final_tasks = load_tasks()
-        if self.task_id in final_tasks:
-            if self.error_message:
-                final_tasks[self.task_id]["status"] = "failed"
-                final_tasks[self.task_id]["error_message"] = self.error_message
-                final_tasks[self.task_id]["speed"] = "0 B/s"
-            else:
-                final_tasks[self.task_id]["status"] = "completed"
-                final_tasks[self.task_id]["progress"] = 100
-                final_tasks[self.task_id]["downloaded_size"] = self.total_size or self.downloaded_size
-                final_tasks[self.task_id]["speed"] = "0 B/s"
-                final_tasks[self.task_id]["eta"] = "已完成"
-
-                # Checksum generation
-                try:
-                    h = hashlib.md5()
-                    with open(self.file_path, 'rb') as f:
-                        for chunk in iter(lambda: f.read(65536), b''):
-                            h.update(chunk)
-                    final_tasks[self.task_id]["md5"] = h.hexdigest()
-                except Exception:
-                    pass
-            save_tasks(final_tasks)
-
-            # Trigger push notification on completed/failed
-            try:
-                # Native Butler notifications via event bus
-                from butler.core.event_bus import event_bus
-                status_zh = "完成" if not self.error_message else "失败"
-                event_bus.emit("NOTIFICATION_PUSH", {
-                    "id": self.task_id,
-                    "title": f"下载任务{status_zh}",
-                    "message": f"文件：{final_tasks[self.task_id]['name']}\n大小：{format_bytes(final_tasks[self.task_id]['total_size'])}",
-                    "category": "downloader"
-                })
-            except Exception:
-                pass
-
-
-# P2P Magnet / ED2K Simulation Task Background Thread
-def run_p2p_simulation(task_id: str, stop_event: threading.Event):
-    """
-    Background simulation worker thread for BT/Magnet and ED2K protocols.
-    Simulates DHT lookup connection, peer discovery, out-of-order slice progression,
-    and beautiful bandwidth graphs.
-    """
-    total_size = 2.4 * 1024 * 1024 * 1024 # 2.4 GB mock movie
-    downloaded = 0
-    start_time = time.time()
-
-    # 1. Look up stage (DHT simulation)
-    dht_nodes = 0
-    while dht_nodes < 256 and not stop_event.is_set():
-        time.sleep(0.3)
-        dht_nodes += 32
-        tasks = load_tasks()
-        if task_id in tasks:
-            tasks[task_id]["status"] = "downloading"
-            tasks[task_id]["speed"] = "DHT 节点查询中..."
-            tasks[task_id]["eta"] = f"已连接 {dht_nodes} 节点"
-            save_tasks(tasks)
-
-    # 2. Main simulated P2P downloads loop
-    peers_count = 12
-    chunk_mb = 18 * 1024 * 1024 # 18 MB/s simulation speed
-
-    while downloaded < total_size and not stop_event.is_set():
-        time.sleep(0.8)
-        # Dynamic speeds fluctuation
-        import random
-        speed_fluct = random.randint(5 * 1024 * 1024, 25 * 1024 * 1024)
-        peers_count = random.randint(18, 54)
-
-        downloaded += speed_fluct
-        if downloaded > total_size:
-            downloaded = total_size
-
-        progress_val = int((downloaded / total_size) * 100)
-
-        tasks = load_tasks()
-        if task_id in tasks:
-            tasks[task_id]["downloaded_size"] = downloaded
-            tasks[task_id]["total_size"] = total_size
-            tasks[task_id]["progress"] = progress_val
-            tasks[task_id]["speed"] = f"{speed_fluct / (1024 * 1024):.1f} MB/s ({peers_count} Peers)"
-            tasks[task_id]["eta"] = calculate_eta(downloaded, total_size, speed_fluct)
-            save_tasks(tasks)
-
-    if not stop_event.is_set():
-        tasks = load_tasks()
-        if task_id in tasks:
-            tasks[task_id]["status"] = "completed"
-            tasks[task_id]["progress"] = 100
-            tasks[task_id]["downloaded_size"] = total_size
-            tasks[task_id]["speed"] = "0 B/s"
-            tasks[task_id]["eta"] = "已完成"
-            save_tasks(tasks)
-
-
-# Scheduler daemon thread checking timing download launches
-def run_scheduler_daemon():
-    """Timer loop checking for scheduled tasks and launching them."""
-    global scheduler_running
-    scheduler_running = True
-    while scheduler_running:
-        time.sleep(10)
-        tasks = load_tasks()
-        now_time = time.strftime("%H:%M")
-
-        launched_any = False
-        for task_id, t in tasks.items():
-            if t.get("status") == "scheduled" and t.get("scheduled_time") == now_time:
-                # Trigger actual download launch
-                t["status"] = "pending"
-                save_tasks(tasks)
-                launch_task_thread(task_id)
-                launched_any = True
-
-        if launched_any:
-            logger.info(f"Scheduler Daemon: Automatically resumed timed downloads matching {now_time}")
-
-
-def launch_task_thread(task_id: str):
-    """Orchestrates actual task execution based on URL protocol."""
-    tasks = load_tasks()
-    task = tasks.get(task_id)
-    if not task:
-        return
-
-    url = task.get("url", "")
-    file_path = task.get("file_path")
-
-    stop_flag = threading.Event()
-
-    # Check protocol
-    if url.startswith("magnet:") or url.startswith("ed2k:"):
-        # Launch BT/ED2K simulation
-        t = threading.Thread(target=run_p2p_simulation, args=(task_id, stop_flag), daemon=True)
-        t.start()
-        with active_lock:
-            ACTIVE_DOWNLOADS[task_id] = {"thread": t, "stop_flag": stop_flag, "manager": None}
-    else:
-        # Launch real multithreaded HTTP/HTTPS downloader
-        limit_speed = 0
-        if task.get("speed_limited"):
-            limit_speed = 200 * 1024 # 200 KB/s limit
-
-        downloader = SegmentedDownloader(
-            task_id=task_id,
-            url=url,
-            file_path=file_path,
-            max_workers=16
-        )
-        downloader.speed_limit = limit_speed
-        downloader.stop_event = stop_flag
-
-        t = threading.Thread(target=downloader.run, daemon=True)
-        t.start()
-        with active_lock:
-            ACTIVE_DOWNLOADS[task_id] = {"thread": t, "stop_flag": stop_flag, "manager": downloader}
-
-
-def stop_active_task(task_id: str):
-    """Stops/pauses an executing task safely."""
-    with active_lock:
-        if task_id in ACTIVE_DOWNLOADS:
-            ACTIVE_DOWNLOADS[task_id]["stop_flag"].set()
-            # If segmented downloader manager exists, stop it
-            mgr = ACTIVE_DOWNLOADS[task_id]["manager"]
-            if mgr:
-                mgr.stop_event.set()
-            ACTIVE_DOWNLOADS.pop(task_id)
-
-
-def execute_network_diagnostics() -> Dict[str, Any]:
-    """Runs a series of real network tests to diagnostic ISP blocking and speeds."""
-    results = {
-        "ping_dns_114": "超时",
-        "ping_dns_google": "超时",
-        "dns_github": "解析失败",
-        "p2p_tracker_blocking": "阻断 (疑似运营商封锁)",
-        "diagnosed_at": time.strftime("%Y-%m-%d %H:%M:%S")
-    }
-
-    # 1. Ping 114 DNS / Google DNS
-    def test_ping(host, port=53):
-        t0 = time.time()
-        try:
-            s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            s.settimeout(1.5)
-            s.connect((host, port))
-            s.close()
-            return f"{int((time.time() - t0)*1000)}ms"
-        except Exception:
-            return "超时"
-
-    results["ping_dns_114"] = test_ping("114.114.114.114")
-    results["ping_dns_google"] = test_ping("8.8.8.8")
-
-    # 2. DNS query
-    try:
-        ip = socket.gethostbyname("github.com")
-        results["dns_github"] = f"正常 ({ip})"
-    except Exception:
-        pass
-
-    # 3. Tracker socket connectivity test (UDP/TCP standard tracker)
-    try:
-        # standard tracker port 80/6969
-        s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        s.settimeout(2.0)
-        s.connect(("tracker.opentrackr.org", 1337))
-        s.close()
-        results["p2p_tracker_blocking"] = "畅通 (支持P2P加速)"
-    except Exception:
-        results["p2p_tracker_blocking"] = "中度封锁 (可能需要设置代理/开启暗网通道)"
-
-    return results
-
-
-# Master entry logic executing request routing from frontend
-def handle_request(action: str, **kwargs) -> Any:
-    # Auto-initialize server threads
-    global scheduler_thread, scheduler_running
-    if not scheduler_running:
-        scheduler_thread = threading.Thread(target=run_scheduler_daemon, daemon=True)
-        scheduler_thread.start()
-        start_streaming_server()
-
-    # Access Jarvis references
-    jarvis_app = kwargs.get("jarvis_app")
-
-    if action == "get_status":
-        return {
-            "status": "ok",
-            "standalone": get_standalone_status(),
-            "download_path": get_downloads_dir(),
-            "config": load_local_config()
-        }
-
-    elif action == "save_settings":
-        download_path = kwargs.get("download_path")
-        if download_path:
-            download_path = download_path.strip()
-            try:
-                os.makedirs(os.path.expanduser(download_path), exist_ok=True)
-            except Exception as e:
-                return {"status": "error", "error": f"无效或无权限的路径: {str(e)}"}
-
-        cfg = load_local_config()
-        if download_path is not None:
-            cfg["download_path"] = download_path
-
-        for k in ["max_threads", "speed_limit", "server_chan", "push_plus"]:
-            if k in kwargs:
-                cfg[k] = kwargs[k]
-
-        save_local_config(cfg)
-        return {"status": "ok", "config": cfg}
-
-    elif action == "list_tasks":
-        return load_tasks()
-
-    elif action == "add_task":
-        url = kwargs.get("url", "").strip()
-        category = kwargs.get("category", "other")
-        scheduled_time = kwargs.get("scheduled_time") # "02:00" string or None
-
-        if not url:
-            return {"error": "链接不能为空"}
-
-        # Protocol conversion (Thunder links decoding)
-        original_url = url
-        is_thunder = False
-        if url.lower().startswith("thunder://"):
-            is_thunder = True
-            try:
-                base64_part = url[10:]
-                decoded = base64_part
-                # Base64 padding correction
-                missing_padding = len(decoded) % 4
-                if missing_padding:
-                    decoded += '=' * (4 - missing_padding)
-                import base64
-                decoded_bytes = base64.b64decode(decoded)
-                # Decode and strip AA/ZZ markers with high resilience
-                decoded_str = decoded_bytes.decode('utf-8', errors='ignore')
-
-                # Align shifted/corrupted protocols
-                if "ttps://" in decoded_str:
-                    decoded_str = decoded_str.replace("ttps://", "https://")
-                elif "ttp://" in decoded_str:
-                    decoded_str = decoded_str.replace("ttp://", "http://")
-
-                match_url = re.search(r'(https?|ftp)://[^\s]+', decoded_str)
-                if match_url:
-                    potential_url = match_url.group(0)
-                    if potential_url.endswith("ZZ"):
-                        potential_url = potential_url[:-2]
-                    url = potential_url
-                elif decoded_str.startswith("AA") and decoded_str.endswith("ZZ"):
-                    url = decoded_str[2:-2]
-            except Exception as e:
-                logger.error(f"Failed to decode Thunder link: {e}")
-
-        # Derive filename
-        name = "unknown_file"
-        if url.startswith("magnet:"):
-            name = "P2P 磁力分享群组"
-            match_dn = re.search(r"dn=([^&]+)", url)
-            if match_dn:
-                name = urllib.parse.unquote(match_dn.group(1))
-        elif url.startswith("ed2k://"):
-            parts = url.split("|")
-            if len(parts) >= 3:
-                name = urllib.parse.unquote(parts[2])
-        else:
-            parsed = urllib.parse.urlparse(url)
-            name = os.path.basename(parsed.path) or "downloaded_file"
-
-        task_id = f"dl_{int(time.time() * 1000)}"
-        file_path = os.path.join(get_downloads_dir(), name)
-
-        tasks = load_tasks()
-        tasks[task_id] = {
-            "id": task_id,
-            "url": url,
-            "original_url": original_url,
-            "name": name,
-            "status": "scheduled" if scheduled_time else "pending",
-            "total_size": 0,
-            "downloaded_size": 0,
-            "progress": 0,
-            "speed": "0 B/s",
-            "eta": "排队中" if not scheduled_time else "已定时",
-            "created_at": time.strftime("%Y-%m-%d %H:%M:%S"),
-            "file_path": file_path,
-            "category": category,
-            "is_p2p": url.startswith("magnet:") or url.startswith("ed2k:"),
-            "scheduled_time": scheduled_time,
-            "speed_limited": False
-        }
-        save_tasks(tasks)
-
-        if not scheduled_time:
-            launch_task_thread(task_id)
-
-        return {"status": "ok", "task_id": task_id, "name": name}
-
-    elif action == "batch_add":
-        urls_text = kwargs.get("urls", "")
-        category = kwargs.get("category", "other")
-        urls_list = [u.strip() for u in urls_text.split("\n") if u.strip()]
-
-        results = []
-        for url in urls_list:
-            res = handle_request("add_task", url=url, category=category)
-            results.append(res)
-        return {"status": "ok", "added": len(results), "results": results}
-
-    elif action == "pause_task":
-        task_id = kwargs.get("task_id")
-        stop_active_task(task_id)
-
-        tasks = load_tasks()
-        if task_id in tasks:
-            tasks[task_id]["status"] = "paused"
-            tasks[task_id]["speed"] = "0 B/s"
-            tasks[task_id]["eta"] = "已暂停"
-            save_tasks(tasks)
-        return {"status": "ok"}
-
-    elif action == "resume_task":
-        task_id = kwargs.get("task_id")
-        tasks = load_tasks()
-        if task_id in tasks:
-            tasks[task_id]["status"] = "pending"
-            tasks[task_id]["eta"] = "开始中..."
-            save_tasks(tasks)
-            launch_task_thread(task_id)
-        return {"status": "ok"}
-
-    elif action == "delete_task":
-        task_id = kwargs.get("task_id")
-        delete_file = kwargs.get("delete_file", False)
-
-        stop_active_task(task_id)
-        tasks = load_tasks()
-        if task_id in tasks:
-            task = tasks.pop(task_id)
-            if delete_file:
-                file_path = task.get("file_path")
-                if file_path and os.path.exists(file_path):
-                    try:
-                        os.remove(file_path)
-                    except Exception:
-                        pass
-            save_tasks(tasks)
-        return {"status": "ok"}
-
-    elif action == "toggle_speed_limit":
-        task_id = kwargs.get("task_id")
-        tasks = load_tasks()
-        if task_id in tasks:
-            tasks[task_id]["speed_limited"] = not tasks[task_id].get("speed_limited", False)
-            save_tasks(tasks)
-
-            # Hot-swap downloader speed limit
-            with active_lock:
-                if task_id in ACTIVE_DOWNLOADS:
-                    mgr = ACTIVE_DOWNLOADS[task_id]["manager"]
-                    if mgr:
-                        limit_val = 200 * 1024 if tasks[task_id]["speed_limited"] else 0
-                        mgr.speed_limit = limit_val
-        return {"status": "ok", "speed_limited": tasks.get(task_id, {}).get("speed_limited", False)}
-
-    elif action == "web_sniff":
-        url = kwargs.get("url", "").strip()
-        if not url:
-            return {"error": "URL不能为空"}
-
-        # Fetch page with headers
-        if not requests:
-            return {"error": "requests 模块未安装，无法进行嗅探"}
-
-        try:
-            r = requests.get(url, timeout=10, headers={
-                "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
-            })
-            html_text = r.text
-
-            links_found = []
-
-            # 1. BeautifulSoup parsing
-            if BeautifulSoup:
-                soup = BeautifulSoup(html_text, "html.parser")
-                # Parse videos tags
-                for video in soup.find_all("video"):
-                    src = video.get("src")
-                    if src:
-                        links_found.append({"title": "HTML5 网页视频", "url": urllib.parse.urljoin(url, src), "type": "video"})
-                    for src_tag in video.find_all("source"):
-                        s_src = src_tag.get("src")
-                        if s_src:
-                            links_found.append({"title": "HTML5 视频源", "url": urllib.parse.urljoin(url, s_src), "type": "video"})
-                # Parse links
-                for a in soup.find_all("a"):
-                    href = a.get("href")
-                    if href:
-                        abs_url = urllib.parse.urljoin(url, href)
-                        text = a.get_text().strip() or "直接媒体下载链接"
-                        if any(abs_url.lower().endswith(ext) for ext in [".mp4", ".mkv", ".avi", ".mp3", ".wav", ".zip", ".rar", ".7z", ".pdf"]):
-                            links_found.append({"title": text[:50], "url": abs_url, "type": "file"})
-            else:
-                # Regex fallback
-                for match in re.findall(r'href=["\'](.*?(\.(mp4|mkv|avi|mp3|zip|rar|7z|pdf)))["\']', html_text, re.IGNORECASE):
-                    abs_url = urllib.parse.urljoin(url, match[0])
-                    links_found.append({"title": os.path.basename(match[0]), "url": abs_url, "type": "file"})
-
-            return {"status": "ok", "links": links_found[:40]}
-        except Exception as e:
-            return {"error": f"网页请求失败: {e}"}
-
-    elif action == "aggregate_search":
-        query = kwargs.get("query", "").strip()
-        if not query:
-            return []
-
-        # Return gorgeous mock results for various domains
-        mock_results = []
-        q_lower = query.lower()
-
-        # 1. Movies Matcher
-        if "毒液" in q_lower or "venom" in q_lower or "电影" in q_lower or "last dance" in q_lower:
-            mock_results.extend([
-                {
-                    "title": "【磁力猫】毒液：最后一舞.Venom.The.Last.Dance.2024.1080p.WEBRip.x264.AAC-HQC",
-                    "size": "2.44 GB",
-                    "url": "magnet:?xt=urn:btih:6fd7a49bb2dcf02410714b627672809a7b97c0f1&dn=Venom.The.Last.Dance.2024.1080p",
-                    "provider": "磁力猫",
-                    "seeds": 142,
-                    "type": "movie"
-                },
-                {
-                    "title": "【电驴先锋】[毒液：最后一舞].Venom.The.Last.Dance.2024.BD1080P.X264.DualAudio-MiniBD.mkv",
-                    "size": "3.81 GB",
-                    "url": "ed2k://|file|[毒液：最后一舞].Venom.The.Last.Dance.2024.BD1080P.mkv|4091280191|7d8a9fb1bc271037748fa0627e8a91cf|/",
-                    "provider": "电驴先锋",
-                    "seeds": 86,
-                    "type": "movie"
-                }
-            ])
-
-        # 2. Software Suite Matcher
-        if "office" in q_lower or "adobe" in q_lower or "软件" in q_lower or "word" in q_lower:
-            mock_results.extend([
-                {
-                    "title": "【小众软件】LibreOffice v24.2.5 简体中文版 (极简开源办公套件)",
-                    "size": "218 MB",
-                    "url": "https://download.documentfoundation.org/libreoffice/stable/24.2.5/win/x86_64/LibreOffice_24.2.5_Win_x86-64.msi",
-                    "provider": "小众软件",
-                    "seeds": 999,
-                    "type": "software"
-                },
-                {
-                    "title": "【小众软件】7-Zip v24.07 简体中文版 (高压缩率工具箱)",
-                    "size": "1.8 MB",
-                    "url": "https://www.7-zip.org/a/7z2407-x64.exe",
-                    "provider": "小众软件",
-                    "seeds": 12450,
-                    "type": "software"
-                }
-            ])
-
-        # 3. Always add default matchers to provide interactive search
-        mock_results.extend([
-            {
-                "title": f"【磁力猫】{query}.Full.Pack.Resources.2024.Zip",
-                "size": "1.22 GB",
-                "url": f"magnet:?xt=urn:btih:3c0cf9ee4f210ea2a5b671a533bf1a989fba7d8a&dn={urllib.parse.quote(query)}.Full.Pack",
-                "provider": "磁力猫",
-                "seeds": 18,
-                "type": "other"
-            },
-            {
-                "title": f"【电驴先锋】[经典合集]{query}.Ultimate.Edition.rar",
-                "size": "850 MB",
-                "url": f"ed2k://|file|{urllib.parse.quote(query)}_Ultimate.rar|891283912|8fba3a2b7cde1203774fa09162ab8de9|/",
-                "provider": "电驴先锋",
-                "seeds": 7,
-                "type": "other"
-            }
-        ])
-
-        return mock_results[:6]
-
-    elif action == "baidu_netdisk_parse":
-        url = kwargs.get("url", "").strip()
-        code = kwargs.get("code", "").strip()
-        if not url:
-            return {"error": "请输入百度网盘分享链接"}
-
-        # Gorgeous simulated extraction showing high-fidelity directory tree
-        time.sleep(1.2) # Elegant extraction delay
-        return {
-            "status": "ok",
-            "title": "【模拟解析】百度网盘：考研数学+数据结构全套教程",
-            "files": [
-                {"name": "01. 考研数学精讲课 - 高等数学.mp4", "size": "854 MB", "is_dir": False},
-                {"name": "02. 考研数学精讲课 - 线性代数.mp4", "size": "612 MB", "is_dir": False},
-                {"name": "03. 数据结构高分指南.pdf", "size": "45.2 MB", "is_dir": False},
-                {"name": "核心讲义源码/", "size": "0 B", "is_dir": True}
-            ]
-        }
-
-    elif action == "network_diagnose":
-        return execute_network_diagnostics()
-
-    elif action == "unzip_extract":
-        task_id = kwargs.get("task_id")
-        passwords = kwargs.get("passwords", [])
-
-        tasks = load_tasks()
-        task = tasks.get(task_id)
-        if not task:
-            return {"error": "任务不存在"}
-
-        file_path = task.get("file_path")
-        if not file_path or not os.path.exists(file_path):
-            return {"error": "下载文件不存在"}
-
-        if not file_path.lower().endswith(".zip"):
-            return {"error": "目前一键密码解压仅支持 .zip 格式文件"}
-
-        # Extract directory named after zip file
-        out_dir = os.path.splitext(file_path)[0]
-        os.makedirs(out_dir, exist_ok=True)
-
-        # Try default standard passwords list
-        common_pwds = ["", "123456", "butler", "admin", "123", "666888"] + passwords
-
-        success = False
-        used_pwd = ""
-        for pwd in common_pwds:
-            try:
-                with zipfile.ZipFile(file_path, "r") as zf:
-                    if pwd:
-                        zf.setpassword(pwd.encode('utf-8'))
-                    zf.extractall(path=out_dir)
-                    success = True
-                    used_pwd = pwd
-                    break
-            except Exception:
-                continue
-
-        if success:
-            return {"status": "ok", "out_dir": out_dir, "password": used_pwd or "无密码"}
-        else:
-            return {"error": "解压密码匹配失败，请手动输入正确的密码"}
-
-    elif action == "media_convert":
-        task_id = kwargs.get("task_id")
-        target_format = kwargs.get("format", "mp3") # mp3, mp4, mkv
-
-        tasks = load_tasks()
-        task = tasks.get(task_id)
-        if not task:
-            return {"error": "任务不存在"}
-
-        file_path = task.get("file_path")
-        if not file_path or not os.path.exists(file_path):
-            return {"error": "文件不存在"}
-
-        # Detect FFMPEG presence
-        ffmpeg_exists = False
-        import shutil
-        if shutil.which("ffmpeg"):
-            ffmpeg_exists = True
-
-        out_path = os.path.splitext(file_path)[0] + f"_converted.{target_format}"
-
-        if ffmpeg_exists:
-            # Real FFMPEG conversion
-            try:
-                import subprocess
-                args = ["ffmpeg", "-y", "-i", file_path]
-                if target_format == "mp3":
-                    args.extend(["-vn", "-acodec", "libmp3lame", "-q:a", "2"])
-                args.append(out_path)
-
-                subprocess.run(args, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-                return {"status": "ok", "out_file": out_path, "method": "FFMPEG 物理转码"}
-            except Exception as e:
-                return {"error": f"转码失败: {e}"}
-        else:
-            # Elegant high-fidelity simulator fallback
-            time.sleep(2.0)
-            try:
-                # Mock create the file
-                with open(out_path, "wb") as f:
-                    f.write(b"MOCK MEDIA CONTENT")
-                return {"status": "ok", "out_file": out_path, "method": "模拟极速云转码 (FFMPEG 未安装)"}
-            except Exception as e:
-                return {"error": str(e)}
-
-    elif action == "push_notification":
-        task_id = kwargs.get("task_id")
-        tasks = load_tasks()
-        task = tasks.get(task_id) if task_id else None
-
-        msg_text = f"下载任务完成：{task['name'] if task else '测试推送'}"
-
-        # Real ServerChan / PushPlus integrations
-        # We look up keys in user configs
-        server_chan_key = kwargs.get("server_chan_key", "")
-        push_plus_key = kwargs.get("push_plus_key", "")
-
-        pushed_channels = []
-        if server_chan_key:
-            try:
-                url = f"https://sctapi.ftqq.com/{server_chan_key}.send"
-                data = urllib.parse.urlencode({"title": "Butler 资源下载器提示", "desp": msg_text}).encode("utf-8")
-                req = urllib.request.Request(url, data=data)
-                urllib.request.urlopen(req, timeout=5)
-                pushed_channels.append("Server酱")
-            except Exception:
-                pass
-
-        if push_plus_key:
-            try:
-                url = "http://www.pushplus.plus/send"
-                payload = json.dumps({
-                    "token": push_plus_key,
-                    "title": "Butler 资源下载器提示",
-                    "content": msg_text
-                }).encode("utf-8")
-                req = urllib.request.Request(url, data=payload, headers={"Content-Type": "application/json"})
-                urllib.request.urlopen(req, timeout=5)
-                pushed_channels.append("PushPlus")
-            except Exception:
-                pass
-
-        return {"status": "ok", "channels": pushed_channels or ["系统托盘通知"]}
-
-    elif action == "get_storage_drives":
-        # Multi-linkage Storage Hub: list all available registered cloud and local storage drives
-        if jarvis_app and hasattr(jarvis_app, "skill_manager"):
-            try:
-                res = jarvis_app.skill_manager.execute("storage_hub", "list_drives")
-                if isinstance(res, dict) and res.get("status") == "ok":
-                    return res.get("drives", [])
-            except Exception as e:
-                logger.error(f"Failed to query Storage Hub drives: {e}")
-
-        # Elegant mock drives if Storage Hub is unconfigured/empty
-        return [
-            {"id": "local_backup", "name": "💻 本地高能库 (Downloads)", "type": "local", "used": 120, "total": 512, "icon": "💻"},
-            {"id": "webdav_alist", "name": "🌐 WebDAV 云盘组 (AList)", "type": "webdav", "used": 540, "total": 1024, "icon": "🌐"},
-            {"id": "onedrive_jules", "name": "☁️ OneDrive 极客盘", "type": "onedrive", "used": 12, "total": 500, "icon": "☁️"}
-        ]
-
-    elif action == "save_to_storage_hub":
-        task_id = kwargs.get("task_id")
-        drive_id = kwargs.get("drive_id", "local_backup")
-        dest_path = kwargs.get("dest_path", "/")
-
-        tasks = load_tasks()
-        task = tasks.get(task_id)
-        if not task:
-            return {"error": "任务不存在"}
-
-        file_path = task.get("file_path")
-        if not file_path or not os.path.exists(file_path):
-            return {"error": "下载文件不可用或已删除"}
-
-        file_name = os.path.basename(file_path)
-        file_size = os.path.getsize(file_path)
-
-        # Real linkage with Storage Hub!
-        if jarvis_app and hasattr(jarvis_app, "skill_manager") and "storage_hub" in jarvis_app.skill_manager.manifests:
-            try:
-                # Copy file to Storage Hub's adapter context or trigger direct transfer/upload
-                # We can physically copy to local path or invoke Storage Hub stream transfer
-                # For high safety: Let's do a physical file relocation or copy, then trigger storage_hub transfer
-                target_dest = os.path.join(PROJECT_ROOT, "data", "storage_hub_transfers", file_name)
-                os.makedirs(os.path.dirname(target_dest), exist_ok=True)
-                shutil.copy2(file_path, target_dest)
-
-                # We can call transfer with this local file
-                # But to keep it simple, robust, and fast: we can move file directly if it's a local adapter,
-                # or trigger the transfer stream in Storage Hub!
-                # Let's return success with physical backup method
-                return {
-                    "status": "ok",
-                    "method": "Storage Hub 共享管道传输",
-                    "destination": f"{drive_id}:{dest_path}/{file_name}",
-                    "size": format_bytes(file_size)
-                }
-            except Exception as e:
-                logger.error(f"Storage Hub sync error: {e}")
-
-        # Interactive High-fidelity simulator upload
-        time.sleep(1.5)
-        return {
-            "status": "ok",
-            "method": "RAM-Pipe 极速内存流管道",
-            "destination": f"{drive_id}:{dest_path}/{file_name}",
-            "size": format_bytes(file_size)
-        }
-
-    elif action == "run":
-        try:
-            import skills.downloader.run as run_module
-            run_module.main()
-            return {"status": "ok"}
-        except Exception as e:
-            return {"error": f"Failed to start standalone downloader: {str(e)}"}
-
-    return {"error": f"Unknown action: {action}"}
-
-
-if __name__ == "__main__":
-    import sys
-    if len(sys.argv) > 1:
-        action = sys.argv[1]
-        # Parse key-value args
-        params = {}
-        for arg in sys.argv[2:]:
-            if "=" in arg:
-                k, v = arg.split("=", 1)
-                params[k] = v
-        print(json.dumps(handle_request(action, **params), ensure_ascii=False, indent=4))
-    else:
-        # Run standalone server blockingly
-        print("=" * 60)
-        print("Butler 资源下载器 - 独立运行模式 (Standalone Server)")
-        print("UI 面板地址: http://localhost:8329/ui/")
-        print("API 接口地址: http://localhost:8329/api/")
-        print("=" * 60)
-        # Start scheduler in daemon thread
-        scheduler_thread = threading.Thread(target=run_scheduler_daemon, daemon=True)
-        scheduler_thread.start()
-        # Start HTTP Server blockingly on port 8329
-        server_address = ("", 8329)
-        try:
-            httpd = http.server.ThreadingHTTPServer(server_address, SafeHTTPRangeHandler)
-            httpd.serve_forever()
-        except KeyboardInterrupt:
-            print("\nStandalone server stopped.")
-        except Exception as e:
-            print(f"Failed to start server: {e}")
+# -*- coding: utf-8 -*-
+"""
+Proxy loader for skills/downloader referencing top-level downloader package.
+"""
+from downloader import (
+    get_standalone_status,
+    get_downloads_dir,
+    SegmentedDownloader,
+    SafeHTTPRangeHandler,
+    start_streaming_server,
+    handle_request,
+    load_tasks,
+    save_tasks,
+    run_scheduler_daemon,
+    launch_task_thread
+)
+
+__all__ = [
+    "get_standalone_status",
+    "get_downloads_dir",
+    "SegmentedDownloader",
+    "SafeHTTPRangeHandler",
+    "start_streaming_server",
+    "handle_request",
+    "load_tasks",
+    "save_tasks",
+    "run_scheduler_daemon",
+    "launch_task_thread"
+]
